changed unworking project
Successfully added device info to DB
