+++ app/src/test/java/com/example/nusra/hit/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.nusra.insoledevice;
+package com.example.nusra.hit;
 
 import org.junit.Test;
